# Back/docs/swagger.yaml
basePath: /
definitions:
...
  DTO.ErrorResponse:
    properties:
      error:
        example: خطایی رخ داده است
        type: string
    type: object
  DTO.LoginRequest:
    properties:
      password:
        maxLength: 100
        minLength: 8
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  DTO.MessageResponse:
    properties:
      message:
        example: عملیات با موفقیت انجام شد
        type: string
    type: object
  DTO.ProfileResponse:
    properties:
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/Models.UserRole'
      section:
        type: string
      status:
        type: boolean
      username:
        type: string
    type: object
  DTO.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  DTO.RegisterRequest:
    properties:
      name:
        maxLength: 20
        minLength: 3
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
      role:
        allOf:
        - $ref: '#/definitions/Models.UserRole'
        description: "\U0001F504 validation جدید"
        enum:
        - admin
        - operator
        - user
        - device
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - name
    - password
    - role
    - username
    type: object
  DTO.ReplaceUserDevicesRequest:
    properties:
      device_id:
        items:
          type: integer
        type: array
    required:
    - device_id
    type: object
...
  DTO.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  DTO.UpdateUserRequest:
    properties:
      is_active:
        type: boolean
      name:
        type: string
      password:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/Models.UserRole'
        enum:
        - admin
        - operator
        - user
        - device
      username:
        type: string
    type: object
...
  DTO.UserDeviceList:
    properties:
      device_id:
        items:
          type: integer
        type: array
    type: object
...
  Models.UserRole:
    enum:
    - admin
    - operator
    - user
    - device
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleOperator
    - RoleUser
    - RoleDevice
info:
  contact: {}
  description: مستندات API پروژه Back (احراز هویت، مدیریت محتوا، مدیریت دستگاه‌ها).
  title: Back API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: احراز هویت با نام کاربری/رمز عبور و دریافت access/refresh token
      parameters:
      - description: اطلاعات ورود
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/DTO.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      summary: ورود کاربر
      tags:
      - Auth
  /auth/logout:
    post:
      description: توکن رفرش کاربر جاری را باطل می‌کند
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: خروج کاربر
      tags:
      - Auth
  /auth/profile:
    get:
      description: اطلاعات کاربر احراز هویت‌شده را برمی‌گرداند
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: پروفایل کاربر جاری
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: با ارسال یک refresh token معتبر، access token و refresh token جدید
        (rotate شده) صادر می‌کند
      parameters:
      - description: توکن رفرش فعلی
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/DTO.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      summary: تمدید توکن
      tags:
      - Auth
  /auth/users:
    get:
      description: لیست تمام کاربران سیستم (فقط ادمین)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/DTO.ProfileResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: لیست کاربران
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: 'در حالت عادی فقط ادمین می‌تواند کاربر (operator/user/device) جدید
        بسازد. استثنا: اگر هنوز هیچ کاربری در سیستم نباشد (اولین راه‌اندازی)، این
        اندپوینت بدون نیاز به توکن قابل استفاده است و کاربر ساخته‌شده اجباراً admin
        می‌شود (مقدار role ارسالی نادیده گرفته می‌شود)'
      parameters:
      - description: اطلاعات کاربر جدید
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/DTO.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/DTO.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      summary: ثبت کاربر جدید
      tags:
      - Users
  /auth/users/{id}:
    delete:
      description: حذف یک کاربر بر اساس شناسه (فقط ادمین)
      parameters:
      - description: شناسه کاربر
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: حذف کاربر
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: ویرایش اطلاعات یک کاربر (فقط ادمین)
      parameters:
      - description: شناسه کاربر
        in: path
        name: id
        required: true
        type: integer
      - description: فیلدهای قابل تغییر
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/DTO.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: به‌روزرسانی کاربر
      tags:
      - Users
  /auth/users/{id}/devices:
    get:
      description: شناسه‌ی دستگاه‌هایی که این کاربر (اپراتور) بهشون دسترسی داره -
        فقط ادمین
      parameters:
      - description: شناسه کاربر
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.UserDeviceList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: لیست دستگاه‌های قابل‌دسترسی یک کاربر
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: جایگزینی کامل لیست دستگاه‌هایی که این کاربر (اپراتور) بهشون دسترسی
        داره - فقط ادمین. برای محدود کردن یک اپراتور (مثلاً یک شرکت پیمانکار) فقط
        به کیوسک‌های خودش
      parameters:
      - description: شناسه کاربر
        in: path
        name: id
        required: true
        type: integer
      - description: لیست شناسه‌ی دستگاه‌ها
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/DTO.ReplaceUserDevicesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DTO.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/DTO.ErrorResponse'
      security:
      - BearerAuth: []
      summary: تنظیم دستگاه‌های قابل‌دسترسی یک کاربر
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: توکن دسترسی رو به‌صورت "Bearer {access_token}" وارد کن.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"